Add `[e] edit PKGBUILD` option when failing to build packages.
